Light Forward Code
Created the code for the Light Forward attack, including the animation for it (still need all the frames, though).
A bug that needs to be fixed is the directional inputs, since currently they're rigged to the forward and backwards keys, which won't work considering how the players swap sides during a match.